requirements.txt
openai
python-dotenv
pandas
pytube
pydub
numpy
yt-dlp
librosa